move ensure_version/ensure_version_or_cargo_install to extra_checks
